refactor: Card 컴포넌트 이동
