+++ src/components/card/index.css.ts
