+++ src/components/index.ts
@@ -2,5 +2,6 @@ import Button from './button';
 import Header from './header';
 import Form from './form';
 import Input from './input';
+import Card from './card';
 
-export { Button, Header, Form, Input };
+export { Button, Header, Form, Input, Card };
