--- src/pages/_components/index.ts
@@ -1,3 +1,0 @@
-import Card from './card';
-
-export { Card };
